+++ source/middleware/generateToken.js
@@ -1,0 +1,7 @@
+import jwt from "jsonwebtoken";
+
+const generateToken = (payload, secret, expiration) => {
+  return jwt.sign(payload, secret, { expiresIn: expiration });
+};
+
+export default generateToken;
